requirements.txt
flask
scikit-learn
nltk
pdfminer.six
gunicorn
